+++ t/apicast-async-reporting.t
@@ -146,7 +146,7 @@ backend client uri: https://127.0.0.1:1953/transactions/authrep.xml?service_toke
 === TEST 3: uses endpoint host as Host header
 when connecting to the backend
 --- main_config
-env RESOLVER=127.0.0.1:1953;
+env RESOLVER=127.0.0.1:$TEST_NGINX_RANDOM_PORT;
 --- http_config
 include $TEST_NGINX_UPSTREAM_CONFIG;
 lua_package_path "$TEST_NGINX_LUA_PATH";
@@ -196,11 +196,11 @@ GET /t?user_key=val
 --- response_body
 all ok
 --- error_code: 200
---- udp_listen: 1953
+--- udp_listen random_port
 --- udp_reply dns
 [ "localhost.example.com", "127.0.0.1", 60 ]
 --- no_error_log
 [error]
---- error_log
-backend client uri: http://localhost.example.com:1984/transactions/authrep.xml?service_token=service-token&service_id=42&usage%5Bhits%5D=2&user_key=val ok: true status: 200
+--- error_log env
+backend client uri: http://localhost.example.com:$TEST_NGINX_SERVER_PORT/transactions/authrep.xml?service_token=service-token&service_id=42&usage%5Bhits%5D=2&user_key=val ok: true status: 200
 --- wait: 3