+++ t/configuration-loading-boot-with-config.t
@@ -15,8 +15,8 @@ __DATA__
 === TEST 1: require configuration file to exist
 should exit when the config file is missing
 --- must_die
---- configuration_file
-t/servroot/html/config.json
+--- configuration_file env
+$TEST_NGINX_SERVER_ROOT/html/config.json
 --- error_log
 config.json: No such file or directory
 --- user_files
@@ -25,8 +25,8 @@ config.json: No such file or directory
 === TEST 2: require valid json file
 should exit when the file has invalid json
 --- must_die
---- configuration_file
-t/servroot/html/config.json
+--- configuration_file env
+$TEST_NGINX_SERVER_ROOT/html/config.json
 --- error_log
 Expected value but found invalid token at character 1
 --- user_files
@@ -35,8 +35,8 @@ not valid json
 
 === TEST 3: empty json file
 should continue as empty json is enough
---- configuration_file
-t/servroot/html/config.json
+--- configuration_file env
+$TEST_NGINX_SERVER_ROOT/html/config.json
 --- request
 GET
 --- error_code: 404